Functional code for shooter motor

+++ src/main/java/frc/robot/Constants.java
@@ -21,18 +21,18 @@ public final class Constants {
     public static final class ShooterConstants {
         public static final int LEAD_SHOOTER_NEO = 5;
         public static final int FOLLOWER_SHOOTER_NEO = 6;
-        public static final int shooter_speed = 1;
+        public static final double shooter_speed = 1.0;
     }
 
     public static final class IntakeConstants {
         public static final int INTAKE_MOTOR = 4;
-        public static final int intake_speed = 1;
+        public static final double intake_speed = 1.0;
     }
 
     public static final class FeederConstants {
         public static final int FEEDER_MOTOR_BOTTOM = 7;
         public static final int FEEDER_MOTOR_TOP = 8;
-        public static final int feeder_speed = 1;
+        public static final double feeder_speed = 1.0;
     }
 
     public static final class ColorWheelConstants {
